examples/requirements.txt
cloudevents==1.6.1
kafka-python==2.0.2
-e ..
